+++ sdk/resources/arm-resources/src/resourceManagementClient.ts
@@ -17,7 +17,10 @@ import {
   _getResourceGroupsOperations,
 } from "./classic/resourceGroups/index.js";
 import { ResourcesOperations, _getResourcesOperations } from "./classic/resources/index.js";
-import { TagsOperations, _getTagsOperations } from "./classic/tags/index.js";
+import {
+  TagsOperationsOperations,
+  _getTagsOperationsOperations,
+} from "./classic/tagsOperations/index.js";
 import { TokenCredential } from "@azure/core-auth";
 import { Pipeline } from "@azure/core-rest-pipeline";
 
@@ -60,7 +63,7 @@ export class ResourceManagementClient {
     this.pipeline = this._client.pipeline;
     this.providerResourceTypes = _getProviderResourceTypesOperations(this._client);
     this.providers = _getProvidersOperations(this._client);
-    this.tags = _getTagsOperations(this._client);
+    this.tagsOperations = _getTagsOperationsOperations(this._client);
     this.resources = _getResourcesOperations(this._client);
     this.resourceGroups = _getResourceGroupsOperations(this._client);
     this.operations = _getOperationsOperations(this._client);
@@ -70,8 +73,8 @@ export class ResourceManagementClient {
   public readonly providerResourceTypes: ProviderResourceTypesOperations;
   /** The operation groups for providers */
   public readonly providers: ProvidersOperations;
-  /** The operation groups for tags */
-  public readonly tags: TagsOperations;
+  /** The operation groups for tagsOperations */
+  public readonly tagsOperations: TagsOperationsOperations;
   /** The operation groups for resources */
   public readonly resources: ResourcesOperations;
   /** The operation groups for resourceGroups */